Further type restructuring
