+++ src/app/store/requests/selectors.ts
@@ -1,5 +1,5 @@
 import Providers from 'app/utilities/Providers';
-import { InitialisedProvider } from "main/providers/types/Provider";
+import { InitialisedProvider } from "main/providers/types";
 import { useCallback, useEffect } from 'react';
 import { useSelector } from 'react-redux';
 import { State, useAppDispatch } from '..';
