+++ src/app/utilities/Providers.ts
@@ -1,5 +1,4 @@
-import { InitOptionalParameters } from 'main/providers/types';
-import { InitialisedProvider } from "main/providers/types/Provider";
+import { InitOptionalParameters , InitialisedProvider } from 'main/providers/types';
 import { ProviderCommands, ProviderEvents } from "main/providers/types/Events";
 import { faFacebookF, faInstagram, faLinkedinIn, faSpotify, IconDefinition } from '@fortawesome/free-brands-svg-icons';
 import { faSquare } from 'app/assets/fa-light';
