+++ src/main/providers/types/Provider.ts
@@ -1,22 +1,10 @@
 import { EmailClient } from 'main/email-client/types';
-import { DataRequestStatus, ProviderFile } from '.';
+import { ProviderFile } from '.';
 
-
-export interface InitialisedProvider {
-    // The key for the provider that supplies the data
-    provider: string;
-    // The account from which the data emanates
-    account?: string;
-    // A URL that is associated with a provider that handles APIs
-    url?: string;
-    // A path- and URL-safe version of the URL
-    hostname?: string;
-    // A random hash which ensures that sessions are kept between various
-    // invocations of browser windows.
-    windowKey: string;
-    status: DataRequestStatus;
-}
-
+/**
+ * The base structure for any provider. A provider is a back-end (e.g. website,
+ * API, organisation, etc.) that can provider data to Aeon.
+ */
 export abstract class Provider {
     protected accountName?: string;
     protected windowKey: string;
@@ -38,6 +26,11 @@ export abstract class Provider {
     }
 }
 
+/**
+ * A specialised form of a Provider that is able to handle Data Requests as
+ * opposed to simple, linear updates. A data request is asynchronous request for
+ * data that may take some time to complete.
+ */
 export interface DataRequestProvider extends Provider {
     /** Dispatch a data request to this Provider. The difference between a
      * regular update and a data request, is that it is asynchronous, and might
@@ -52,20 +45,44 @@ export interface DataRequestProvider extends Provider {
     parseDataRequest?(extractionPath: string, identifier?: string | number): Promise<ProviderFile[]>;
 }
 
+/**
+ * A specialised form of a Provider that is able to handle Data Requests as
+ * opposed to simple, linear updates. A data request is asynchronous request for
+ * data that may take some time to complete.
+ */
 export abstract class DataRequestProvider extends Provider {
     /** The amount of days that are required between successive data requests */
     public static dataRequestIntervalDays: number;
 }
 
+/**
+ * A specialised form of a DataRequestProvider that requires email in order to
+ * operate. This means concretely, that a user will be required to link an email
+ * account to this provider, which is then accessible in the class.
+ */
 export abstract class EmailDataRequestProvider extends DataRequestProvider {
+    /* An email client that is available for use in this provider. This will
+    automatically be set by Aeon when constructing the class. */
     protected email: EmailClient;
     setEmailClient(email: EmailClient): void {
         this.email = email;
     }
 }
 
+/**
+ * A specialised form of a DataRequestProvider that implements the Open Data
+ * Rights API. This practically means that users will be required to enter a URL
+ * when creating an instance of this provider, which is then available for use
+ * in the class itself.
+ * NOTE: This type is used primarily in an implemeting class called
+ * OpenDataRights (see ../open-data-rights). You don't need to create a seperate
+ * class for each organisation implementing an Open Data Rights API.
+ * @see https://whitepaper.open-data-rights.org
+ */
 export abstract class OpenDataRightsProvider extends DataRequestProvider {
+    /* The Open Data Rights API URL, for use in the class */
     protected url: string;
+    /* A convenience method to pass to withSecureWindow */
     protected windowParams: {
         key: string;
         origin: string;
