+++ src/main/providers/types/index.ts
@@ -14,6 +14,21 @@ export interface DataRequestStatus {
     requestId?: string | number;
 }
 
+export interface InitialisedProvider {
+    // The key for the provider that supplies the data
+    provider: string;
+    // The account from which the data emanates
+    account?: string;
+    // A URL that is associated with a provider that handles APIs
+    url?: string;
+    // A path- and URL-safe version of the URL
+    hostname?: string;
+    // A random hash which ensures that sessions are kept between various
+    // invocations of browser windows.
+    windowKey: string;
+    status: DataRequestStatus;
+}
+
 export enum ProviderUpdateType {
     UPDATE = 'update',
     DATA_REQUEST = 'data_request'
@@ -23,3 +38,4 @@ export type InitOptionalParameters = {
     accountName?: string;
     apiUrl?: string;
 }
+
